% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-pkgs.R
\name{check_pkgs_availability}
\alias{check_pkgs_availability}
\title{Check Packages Availability}
\usage{
check_pkgs_availability(pkgs)
}
\arguments{
\item{pkgs}{Character vector with package names checked.}
}
\value{
Named logical indicating whether each package is available on the
machine.
}
\description{
See example for common way would be used in {indexrfiles}.
}
\examples{
library(indexrfiles)
library(dplyr)

file_lines <- "
library(dplyr)
require(tidyr)
library(madeUpPkg)

as_tibble(mpg) \%>\%
  group_by(class) \%>\%
  nest() \%>\%
  mutate(stats = purrr::map(data,
                            ~lm(cty ~ hwy, data = .x)))

made_up_fun()
"

file_output <- tempfile(fileext = ".R")
writeLines(file_lines, file_output)

get_pkgs(file_output) \%>\%
  check_pkgs_availability()

}
