% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-files-helpers.R
\name{did_safely_error}
\alias{did_safely_error}
\title{Did Safely Error}
\usage{
did_safely_error(safely_named_list)
}
\arguments{
\item{safely_named_list}{Named list as outputted by \code{purrr::safely()} where
each index contains an "error" and a "result" element.}
}
\value{
logical vector
}
\description{
Did Safely Error
}
