% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funspotr.R
\docType{package}
\name{funspotr-package}
\alias{funspotr}
\alias{funspotr-package}
\title{funspotr: Identify Functions & Packages in Files}
\description{
To learn more about funspotr see https://github.com/brshallo/funspotr
}
\seealso{
Useful links:
\itemize{
  \item \url{https://brshallo.github.io/funspotr/}
  \item Report bugs at \url{https://github.com/brshallo/funspotr/issues}
}

}
\author{
\strong{Maintainer}: Bryan Shalloway \email{brshallodev@gmail.com}

}
\keyword{internal}
