% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-pkgs.R
\name{get_pkgs_from_DESCRIPTION}
\alias{get_pkgs_from_DESCRIPTION}
\title{Get package dependencies from DESCRIPTION file}
\usage{
get_pkgs_from_DESCRIPTION(DESCRIPTION_path)
}
\arguments{
\item{DESCRIPTION_path}{Path to DESCRIPTION file}
}
\value{
Character vector of packages.
}
\description{
Given explicit path to DESCRIPTION file return package dependencies therein.
inspiration: https://stackoverflow.com/a/30225680/9059865
}
\examples{
library(funspotr)
get_pkgs_from_DESCRIPTION(
  "https://raw.githubusercontent.com/brshallo/animatrixr/master/DESCRIPTION"
)
}
