% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-files-gh-repo.R
\name{github_contents}
\alias{github_contents}
\title{Github Contents}
\usage{
github_contents(repo, branch = NULL)
}
\arguments{
\item{repo}{Github user/repo}

\item{branch}{Default is \code{NULL} which will test "main" and then if that is
invalid, "master".}
}
\value{
Character vector of contents from github repo.
}
\description{
List files and folders in a github repo.
inspiration: https://stackoverflow.com/a/25485782/9059865
}
