% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-files-gh-repo.R
\name{github_contents_urls}
\alias{github_contents_urls}
\title{Github Contents and URLS}
\usage{
github_contents_urls(repo, branch = "main")
}
\arguments{
\item{repo}{Github user/repo}

\item{branch}{Default is \code{NULL} which will test "main" and then if that is
invalid, "master".}
}
\value{
Dataframe with columns \code{relative_paths} and \code{absolute_paths}.
}
\description{
Runs \code{funspotr:::github_contents()} and then maps this through
\code{github_rbf_to_url()} returning URL's to the "raw" location.
}
