% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-gists.R
\name{github_gists}
\alias{github_gists}
\title{Get GitHub Gist Urls}
\usage{
github_gists(user)
}
\arguments{
\item{user}{Character.}
}
\value{
Dataframe with \code{contents} and \code{urls} of file paths.
}
\description{
Given a username, return a dataframe of all the gists by that user. Note that
(similar to other \verb{github_*} functions) if run repeatedly may error due to
hitting github API rate limits.
}
\details{
Returned output does not filter to only R and Rmarkdown files by default, see
example for typical use with \code{funspotr:::str_detect_r_rmd()} -- I use a lot
of .md files for gists so should probably write something to parse out the R
code chunks from .md files.
}
\examples{
library(dplyr)
library(funspotr)

# pulling and analyzing my R file github gists
gists_urls <- github_gists("brshallo") \%>\%
  filter(funspotr:::str_detect_r_rmd(contents))

gists_urls

# Will just parse the first 2 files/gists
contents <-
  github_spot_funs(custom_urls = slice(gists_urls, 1:2))

contents \%>\%
  unnest_github_results()
}
