% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexrfiles-package.R
\docType{package}
\name{indexrfiles-package}
\alias{indexrfiles}
\alias{indexrfiles-package}
\title{indexrfiles: Identify Functions & Packages in Files}
\description{
To learn more about indexrfiles see https://github.com/brshallo/indexrfiles
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/tidyverse/dplyr/issues}
}

}
\author{
\strong{Maintainer}: Bryan Shalloway \email{brshallodev@gmail.com}

}
\keyword{internal}
