% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-pkgs.R
\name{install_missing_pkgs}
\alias{install_missing_pkgs}
\title{Install missing packages}
\usage{
install_missing_pkgs(pkgs_availability)
}
\arguments{
\item{pkgs_availability}{Named logical vector where names are packages --
generally the output of running \code{check_pkgs_availability()}.}
}
\value{
Installs packages from cran using \code{remotes::install_cran()} if
available else \code{install.packages()}
}
\description{
Attempt to install missing packages from CRAN. This whole set of functions
would be better if handled by taking advantage of something like
\code{renv::dependencies()}.
}
