% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot-funs.R
\name{list_functions_in_file}
\alias{list_functions_in_file}
\title{List Functions in File}
\usage{
list_functions_in_file(file_path, show_each_use = FALSE)
}
\arguments{
\item{show_each_use}{If changed to \code{TRUE} will return each instance a function
is used rather than once for everything}
}
\description{
Simply a wrapper on \code{utils::getParseData()} . Is mostly copied from:
NCmisc::list.functions.in.file(). Rewrote because function made order of
output dependent on packages. Also wanted to add \code{show_each_use} arg.
}
