% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot-tags.R
\name{spot_tags}
\alias{spot_tags}
\title{Spot Tags}
\usage{
spot_tags(file_path = knitr::current_input(), used = FALSE, drop_knitr = FALSE)
}
\arguments{
\item{file_path}{Default is the file being knitted but can change to some
other file (e.g. in cases where the code for the post may reside in a
different file).}

\item{used}{Default is \code{FALSE}. If \code{TRUE} will pass to \code{show_pkgs_used()}
rather than \code{show_pkgs()}. (Mainly useful for showing actual packages used
rather than meta-packages being called like \code{tidyverse} or \code{tidymodels}).}

\item{drop_knitr}{Many blogdown posts have \code{knitr::opts_chunk$set()} in them
and you may not want this tag showing-up. Default is to keep this, but set
to \code{FALSE} to drop "knitr" from being tagged.

... a better approach may be to just skip the "setup" chunk during parsing...}
}
\value{
String in a format meant to be inserted in tags argument of YAML header.
}
\description{
Put function in your blogdown post's YAML header to have the packages be the
packages used in your post (wrapper around \code{funspotr::spot_pkgs()}).
}
\details{
\preformatted{tags: ["`r funspotr::spot_tags()`"]
}

Thanks Yihui for getting this working and for suggesting the function! Note
requires blogdown >= 1.9 to work
(https://github.com/rstudio/blogdown/issues/647).
}
\examples{

# Put this in your blogdown posts YAML header to autogenerate tags based on pkgs
# tags: ["`r funspotr::spot_tags()`"]
}
\seealso{
show_pkgs show_pkgs_used
}
