% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-spot.R
\name{unnest_github_results}
\alias{unnest_github_results}
\title{Unnest Github Results}
\usage{
unnest_github_results(df)
}
\arguments{
\item{df}{Dataframe outputted by \verb{github_spot*()} that contains a \code{spotted} list-column.}
}
\value{
An unnested dataframe
}
\description{
After running \code{github_spot()}
}
\examples{
library(funspotr)
library(dplyr)

github_spot_funs("brshallo/feat-eng-lags-presentation", branch = "main") \%>\%
  unnest_github_results()
}
