% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-get.R
\name{github_contents}
\alias{github_contents}
\title{Github Contents}
\usage{
github_contents(repo, branch = "main")
}
\arguments{
\item{repo}{Github user/repo}

\item{branch}{Default is "main", "master" is likely also common.}
}
\value{
Character vector of contents from github repo.
}
\description{
List files and folders in a github repo.
inspiration: https://stackoverflow.com/a/25485782/9059865
}
\examples{

funspotr:::github_contents("brshallo/feat-eng-lags-presentation", branch = "main")
}
