% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-spot.R
\name{github_contents_urls}
\alias{github_contents_urls}
\title{Github Contents and URLS}
\usage{
github_contents_urls(repo, branch = "main")
}
\arguments{
\item{repo}{Github user/repo}

\item{branch}{Default is "main"; "master" is also common.}
}
\value{
Dataframe with columns \code{contents} and \code{urls}.
}
\description{
Runs \code{funspotr:::github_contents()} and then maps this through
\code{github_rbf_to_url()} returning URL's to the "raw" location.
}
\examples{

funspotr:::github_contents_urls("brshallo/feat-eng-lags-presentation", branch = "main")
}
