% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-files-gh-gists.R
\name{list_files_github_gists}
\alias{list_files_github_gists}
\title{List Files in GitHub Gists}
\usage{
list_files_github_gists(
  user,
  pattern = stringr::regex("(r|rmd|rmarkdown|qmd)$", ignore_case = TRUE)
)
}
\arguments{
\item{user}{GitHub username, e.g. "brshallo"}

\item{pattern}{Regex pattern to keep only matching files. Default is
\code{stringr::regex("(r|rmd|rmarkdown|qmd)$", ignore_case = TRUE)} which will
keep only R, Rmarkdown and Quarto documents. To keep all files use \code{"."}.}
}
\value{
Dataframe with columns of \code{relative_paths} and \code{absolute_paths} for
file path locations. \code{absolute_paths} will be urls to raw files.
}
\description{
Return a dataframe containing the paths of files in a GitHub user's gists.
Generally used prior to \verb{spot_\{funs/pkgs\}_files()}.
}
\examples{
\donttest{
library(dplyr)
library(funspotr)

# pulling and analyzing R file github gists
gh_urls <- list_files_github_gists("brshallo")

# Will just parse the first 2 files/gists
contents <- spot_funs_files(slice(gh_urls, 1:2))

contents \%>\%
  unnest_results()
}
}
\seealso{
list_files_github_repo, list_files_wd
}
