% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot-funs.R
\name{list_functions_in_file_to_df}
\alias{list_functions_in_file_to_df}
\title{List Functions in File to Dataframe}
\usage{
list_functions_in_file_to_df(funs)
}
\arguments{
\item{funs}{List output returned from running
NCmisc::list.functions.in.file()}
}
\description{
This is a helper function that converts output from
\code{NCmisc::list.functions.in.file()} into a tibble and cleans-up a few things.
}
\details{
\code{list_functions_in_file_to_df()} is the last step inside \code{spot_funs()} --
check there for documentation on the returned output.
}
