% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-plot.R
\name{network_plot}
\alias{network_plot}
\title{Funspotr Network Plot}
\usage{
network_plot(df, to = pkgs, show_each_use = FALSE)
}
\arguments{
\item{df}{Dataframe containing columns \code{contents}, \code{urls} and either \code{funs}
or \code{pkgs}. Generally the output from running:
\verb{github_spot_*() \%>\% unnest_github_results()}}

\item{to}{\code{funs} or \code{pkgs}}

\item{show_each_use}{Binary, default is \code{FALSE}. If \code{TRUE} edge thickness
will be based on the number of times a package or function is used.}
}
\value{
visNetwork plot
}
\description{
Output simple network plot using
\href{https://github.com/datastorm-open/visNetwork}{visNetwork} connecting either
\code{funs} or \code{pkgs} to \code{contents}/\code{urls}.
}
\examples{
library(dplyr)
library(funspotr)

gh_ex_pkgs <- github_spot_pkgs(
  repo = "brshallo/feat-eng-lags-presentation",
  branch = "main")

gh_ex_pkgs \%>\%
  unnest_github_results() \%>\%
  network_plot(to = pkgs)

}
