% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot-pkgs.R
\name{spot_pkgs_from_description}
\alias{spot_pkgs_from_description}
\title{Spot package dependencies from DESCRIPTION file}
\usage{
spot_pkgs_from_description(DESCRIPTION_path)
}
\arguments{
\item{DESCRIPTION_path}{Path to DESCRIPTION file}
}
\value{
Character vector of packages.
}
\description{
Given explicit path to DESCRIPTION file return package dependencies therein.
Inspiration: \href{https://stackoverflow.com/a/30225680/9059865}{blogdown#647}.
}
\examples{
funspotr::spot_pkgs_from_description(
  "https://raw.githubusercontent.com/brshallo/animatrixr/master/DESCRIPTION"
)
}
\keyword{internal}
