% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-get.R
\name{str_detect_r_rmd}
\alias{str_detect_r_rmd}
\title{Keep only R and Rmarkdown files}
\usage{
str_detect_r_rmd(contents, rmv_index = TRUE)
}
\arguments{
\item{contents}{Vector of folder and file pathways.}

\item{rmv_index}{Logical, most repos containing blogdown sites will have an
index.R file. By default these are removed from list of files.}
}
\value{
Subsetted character vector.
}
\description{
Keep only R and Rmarkdown files
}
\examples{
library(dplyr)

funspotr:::github_contents("brshallo/feat-eng-lags-presentation", branch = "main") \%>\%
  funspotr:::str_detect_r_rmd()
}
