% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-files-gh-repo.R
\name{str_detect_r_rmd}
\alias{str_detect_r_rmd}
\title{String Detect R or Rmarkdown File endings}
\usage{
str_detect_r_rmd(contents, rmv_index = TRUE)
}
\arguments{
\item{contents}{Character vector of file path.}

\item{rmv_index}{Logical, default to \code{TRUE}, most repos containing blogdown
sites will have an index.R file at the root. Change to \code{FALSE} if you don't
want this file removed.}
}
\value{
Logical vector.
}
\description{
Return \code{TRUE} for only R and Rmarkdown files, else \code{FALSE}.
}
\examples{
files <- c("file1.R", "file2.Rmd", "file3.Rmarkdown", "file4.Rproj")
funspotr:::str_detect_r_rmd(files)
}
